break up resource into params.rs + fixes
- PostParams now can take field manager fields to reflecte CreateOptions
- DeleteParams::dry_run is back to bool to reflect the rest

It's weird how the api accepts only the single string "All" inside a
map, or nothing. So made a small serializer for it.
